django/requirements.txt
amqp==5.0.9
billiard
celery
Django==3.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-versatileimagefield==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
kombu
Pillow
psycopg2-binary==2.8.6
pyfcm==1.5.1
PyJWT==2.0.1
python-dateutil==2.7.5
python-decouple==3.1
python-magic==0.4.18
python-magic-bin
pytz==2021.1
requests
six==1.15.0
sqlparse==0.4.1
uWSGI
vine==5.0.0